Fix tests. Increase Pythonicness
A mock function was added to the tests to allow testing without a physical
Raspberry Pi or Camera. (Hello Travis...)

Initial steps towards Pythonicness were attempted.
These include the addition of __init__.py files, and the renaming of the
`src` directory to `pimotion`.
Renamed src directory to pimotion directory to be more Pythonic.